refactoring #2

+++ include/animation/animator.hpp
@@ -5,5 +5,5 @@
 
 namespace Animation
 {
-    void move(WINDOW* win, Position startPosition, Position endPosition, const char* file, double seconds, unsigned short color);
+    void move(WINDOW* win, Position startPosition, Position endPosition, const char* file, double seconds, unsigned color);
 }
