+++ src/scenes/game_scene.cpp
@@ -14,16 +14,14 @@ void GameScene::load()
         (yMax / 2) - BOARD_SPACE_Y,
         (xMax / 2) + BOARD_SPACE_X
     ));
-    // dialog = new Dialog(Position((yMax / 2) + 5, (xMax / 2)), 
-    // L"for move: ← → ↑ ↓ | for rotate: r | for set: enter");
     wrefresh(sceneWin);
     refresh();
 }
 
-void GameScene::endMenu(const char* file, unsigned short color)
+void GameScene::endMenu(const char* file, unsigned color)
 {
     wclear(sceneWin);
     Position fileSize = getSize(file);
-    Position winPosition {(short)(yMax / 2 - fileSize.y / 2), (short)(xMax / 2 - fileSize.x / 2)};
+    Position winPosition {yMax / 2 - fileSize.y / 2, xMax / 2 - fileSize.x / 2};
     printImage(sceneWin, winPosition, file, color);
 }
